% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_aboutTab.R
\name{aboutTab}
\alias{aboutTab}
\title{About tab server.}
\usage{
aboutTab(input, output, session)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\description{
About tab server.
}
