% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aboutTab.R
\name{aboutTabUI}
\alias{aboutTabUI}
\title{About tab user interface.}
\usage{
aboutTabUI(id)
}
\arguments{
\item{id}{namespace id}
}
\value{
the about tab panel
}
\description{
About tab user interface.
}
