% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtilities.R
\name{addNowLabelAndIntercept}
\alias{addNowLabelAndIntercept}
\title{Add now label ('Past' and 'Future') and X intercept to an existing plot.}
\usage{
addNowLabelAndIntercept(plot, now)
}
\arguments{
\item{plot}{a plot}

\item{now}{now date}
}
\value{
a new plot with the X intercept and 'past' and 'future' labels
}
\description{
Add now label ('Past' and 'Future') and X intercept to an existing plot.
}
