% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.R
\name{addPatient}
\alias{addPatient}
\title{Add a patient into the database.}
\usage{
addPatient(patient)
}
\arguments{
\item{patient}{patient}
}
\value{
the patient ID
}
\description{
Add a patient into the database.
}
