% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTab.R
\name{applyModel}
\alias{applyModel}
\title{Apply new selected model.
The old model is replaced by the new one.
If some covariates are missing, they are added automatically into the covariates data frame, with their default value.}
\usage{
applyModel(val)
}
\arguments{
\item{val}{main reactive container}
}
\description{
Apply new selected model.
The old model is replaced by the new one.
If some covariates are missing, they are added automatically into the covariates data frame, with their default value.
}
