% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelLibrary.R
\docType{data}
\name{bergmann2014_base}
\alias{bergmann2014_base}
\title{bergmann2014_base model.}
\format{tdmore model}
\usage{
bergmann2014_base
}
\description{
bergmann2014_base model.
}
\keyword{datasets}
