% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelLibrary.R
\docType{data}
\name{bergmann2014_mixture}
\alias{bergmann2014_mixture}
\title{bergmann2014_mixture model.}
\format{tdmore mixture model}
\usage{
bergmann2014_mixture
}
\description{
bergmann2014_mixture model.
}
\keyword{datasets}
