% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{bergmann2014_to_delete}
\alias{bergmann2014_to_delete}
\title{Temporary model.
Will be deleted.}
\format{RxODE model}
\usage{
bergmann2014_to_delete
}
\description{
Temporary model.
Will be deleted.
}
\keyword{datasets}
