% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{calculation}
\alias{calculation}
\title{This module performs all of the required calculation for the application.}
\usage{
calculation(state)
}
\description{
Call it in your application using `calculation(state)`
}
\section{Output}{

The module sets the following state variables:
* `populationPredict` a data-frame with prediction data on the population level
* `fit` a tdmorefit object
* `individualPredict` a data-frame with prediction data for the individual
* `recommendation` a data.frame in the same form as `regimen`, but where the AMT column was adapted
}

