% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistencyCheck.R
\name{checkCovariates}
\alias{checkCovariates}
\title{Check covariates.}
\usage{
checkCovariates(covariates)
}
\arguments{
\item{covariates}{named numeric with the covariates}
}
\description{
Check covariates.
}
