% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_mongodb.R
\name{composeUrl}
\alias{composeUrl}
\title{composeUrl}
\arguments{
\item{host}{hostname}

\item{user}{username, or NULL for no user}

\item{password}{password, or NULL for no password}
}
\value{
composes a url of mongodb://user:password@host
}
\description{
This is a function to easily compose a URL for MongoDB
}
