% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtilities.R
\name{convertDataToTdmore}
\alias{convertDataToTdmore}
\title{Convert main data to TDMore domain.}
\usage{
convertDataToTdmore(doses, obs, now)
}
\arguments{
\item{doses}{doses}

\item{obs}{observations}

\item{now}{now date, POSIXlt date}
}
\value{
to be described
}
\description{
Convert main data to TDMore domain.
}
