% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientModel.R
\name{covariateModelToJson}
\alias{covariateModelToJson}
\title{Convert covariate model to JSON structure.}
\usage{
covariateModelToJson(covariateModel)
}
\arguments{
\item{covariateModel}{the covariate model}
}
\value{
data frame with date column (full date), covariate name and value
}
\description{
Convert covariate model to JSON structure.
}
