% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_hotTable.R
\name{covariatesTable}
\alias{covariatesTable}
\title{Create a user interface to represent the covariates}
\usage{
covariatesTable(input, output, session, state)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{state}{\code{reactiveValues} object with at least the values \code{now}, \code{observed} and \code{model}}
}
\description{
The interface consists of a \emph{table} (rhandsontable) that reads from the \code{state$covs} data.frame. It has at least two columns:
}
\details{
\enumerate{
\item \code{date} is a POSIXct date that is represented as YYYY/MM/DD in the table
\item \code{time} is a character string HH:MM that is represented as a dropdown box
\item and further: columns for each covariate in the model
}

The table has a horizontal line to reflect where \code{state$now} is situated. If \code{state$now} is empty, the line is not shown.

Edits to this observed table are automatically synchronized back to \code{state$covs}. The element ensures the rows are chronological.

The user interface also features a button to \strong{add} new covariates. Adding to an empty table creates a line for an observation at \strong{state$now}.
Adding to an existing table duplicates the last row. The time is shifted by \verb{24 hours}.
}
