% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataConversion.R
\name{covsToTdmore}
\alias{covsToTdmore}
\title{Covariates conversion (shinyTDMore -> TDMore).}
\usage{
covsToTdmore(covs, firstDoseDate, model)
}
\arguments{
\item{covs}{shinyTDMore covariates}

\item{firstDoseDate}{first dose date}

\item{model}{model}
}
\value{
TDMore covariates
}
\description{
Covariates conversion (shinyTDMore -> TDMore).
}
