% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{createDBConfig}
\alias{createDBConfig}
\title{Create a DB configuration.}
\usage{
createDBConfig(collection, db, user = NULL, password = NULL,
  server = NULL)
}
\arguments{
\item{collection}{a mongo collection, not null}

\item{db}{a mongo collection, not null}

\item{user}{a mongo collection}

\item{password}{a mongo collection}

\item{server}{a mongo collection}
}
\value{
a shinytdmore_db_config object
}
\description{
Create a DB configuration.
}
