% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.R
\name{createPatient}
\alias{createPatient}
\title{Patient constructor.}
\usage{
createPatient(firstname, lastname)
}
\arguments{
\item{firstname}{patient's firstname}

\item{lastname}{patient's lastname}
}
\value{
the patient
}
\description{
Patient constructor.
}
