% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{D0_14_Troughs_1cpt_BASE_TSFD_MPC}
\alias{D0_14_Troughs_1cpt_BASE_TSFD_MPC}
\title{MPC Tacro model - 1 cpt.}
\format{RxODE model}
\usage{
D0_14_Troughs_1cpt_BASE_TSFD_MPC
}
\description{
MPC Tacro model - 1 cpt.
}
\keyword{datasets}
