% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{D7_AUC_2cpt_Tlag_CYP3A5_alloWT}
\alias{D7_AUC_2cpt_Tlag_CYP3A5_alloWT}
\title{Tacro model - 2 cpt.}
\format{RxODE model}
\usage{
D7_AUC_2cpt_Tlag_CYP3A5_alloWT
}
\description{
Tacro model - 2 cpt.
}
\keyword{datasets}
