% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{D7_AUC_2cpt_Tlag_CYP3A5_alloWT_MPC}
\alias{D7_AUC_2cpt_Tlag_CYP3A5_alloWT_MPC}
\title{MPC Tacro model - 2 cpt.}
\format{RxODE model}
\usage{
D7_AUC_2cpt_Tlag_CYP3A5_alloWT_MPC
}
\description{
MPC Tacro model - 2 cpt.
}
\keyword{datasets}
