% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\docType{class}
\name{Database}
\alias{Database}
\title{Database Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Database
}
\description{
Database class description
}
\details{
Database class
}
\keyword{datasets}
