% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveProject.R
\name{dataHasChanged}
\alias{dataHasChanged}
\title{Say if data (measures, doses, covariates, now date or model) have been changed by the user.}
\usage{
dataHasChanged(val)
}
\arguments{
\item{val}{main reactive container}
}
\value{
a logical value
}
\description{
Say if data (measures, doses, covariates, now date or model) have been changed by the user.
}
