% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtilities.R
\name{dateAndTimeToPOSIX}
\alias{dateAndTimeToPOSIX}
\title{Convert date and time vectors to POSIX.}
\usage{
dateAndTimeToPOSIX(date, time)
}
\arguments{
\item{date}{date vector}

\item{time}{character vector with '\%H:\%M'-formatted times}
}
\description{
Convert date and time vectors to POSIX.
}
