% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelLibrary.R
\docType{data}
\name{default_alloWT_MPC}
\alias{default_alloWT_MPC}
\title{default_alloWT_MPC model.}
\format{tdmore model}
\usage{
default_alloWT_MPC
}
\description{
default_alloWT_MPC model.
}
\keyword{datasets}
