% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultApp.R
\name{defaultApp}
\alias{defaultApp}
\title{Default full-featured application}
\usage{
defaultApp(...)
}
\arguments{
\item{...}{default values for the application (e.g. `regimen`, `observed`, etc)}
}
\description{
Default full-featured application
}
