% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{defaultDBConfig}
\alias{defaultDBConfig}
\title{Get the default DB configuration.}
\usage{
defaultDBConfig()
}
\value{
a shinytdmore_db_config object
}
\description{
Get the default DB configuration.
}
