% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientModel.R
\name{doseModelToJson}
\alias{doseModelToJson}
\title{Convert dose model to JSON structure.}
\usage{
doseModelToJson(doseModel)
}
\arguments{
\item{doseModel}{the dose model}
}
\value{
data frame with date column (full date) and amount column
}
\description{
Convert dose model to JSON structure.
}
