% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_hotTable.R
\name{doseTable}
\alias{doseTable}
\title{Server component for doseTable}
\usage{
doseTable(input, output, session, state)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{state}{`reactiveValues` object with the values `now`, `regimen` and `model`.
`now` can be absent, in which case the `now` time is not shown graphically in the table.
Also, the addDose command will then not use the current date by default.

`model` should at least contain formulation metadata. If not, the corresponding reactives
that set the output doseTable will show a silent shiny error (see [shiny::req()] for more information).}
}
\description{
Server component for doseTable
}
