% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xEditableInput.R
\name{editableCombodate}
\alias{editableCombodate}
\title{Create an in-place editable combodate. This function adds some default options for the format.}
\usage{
editableCombodate(inputId, value = "", options = list())
}
\arguments{
\item{inputId}{The `input` slot that will be used to access the value.}

\item{value}{Initial value.}

\item{options}{A list of general and input-type-dependent options for
[X-editable](http://github.com/vitalets/x-editable). Details
[here](http://vitalets.github.io/x-editable/docs.html#editable).}
}
\description{
Create an in-place editable combodate. This function adds some default options for the format.
}
\details{
This function adds a default format YYYY-MM-DD HH:mm, and also sets the template and viewformat arguments to the same value.
}
