% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xEditableInput.R
\name{editableHtml5}
\alias{editableHtml5}
\title{Create an in-place editable html5 element.
This creates an X-editable of type `text`, with the appropriate template for the input element.}
\usage{
editableHtml5(inputId, type = "text", options = list())
}
\arguments{
\item{inputId}{The `input` slot that will be used to access the value.}

\item{type}{an allowed html5 [input type](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input)}

\item{options}{A list of general and input-type-dependent options for
[X-editable](http://github.com/vitalets/x-editable). Details
[here](http://vitalets.github.io/x-editable/docs.html#editable).}
}
\description{
Create an in-place editable html5 element.
This creates an X-editable of type `text`, with the appropriate template for the input element.
}
