% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientForm.R
\name{eraseCovariateForm}
\alias{eraseCovariateForm}
\title{Erase the covariate form.}
\usage{
eraseCovariateForm(session)
}
\arguments{
\item{session}{the shiny session object}
}
\description{
Erase the covariate form.
}
