% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{estimate}
\alias{estimate}
\title{Link to protected `tdmore:::estimate` function}
\usage{
estimate(...)
}
\arguments{
\item{...}{passed to estimate method}
}
\value{
a tdmorefit object
}
\description{
Link to protected `tdmore:::estimate` function
}
