% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientModel.R
\name{exportPatientToJsonFile}
\alias{exportPatientToJsonFile}
\title{Export patient to JSON file.}
\usage{
exportPatientToJsonFile(filePath, patientModel)
}
\arguments{
\item{filePath}{the file path (including file name)}

\item{patientModel}{the patient model to be exported}
}
\description{
Export patient to JSON file.
}
