% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_fitPlot.R
\name{fitPlot}
\alias{fitPlot}
\title{Fitplot creates a Population, Fit and Recommendation plot}
\usage{
fitPlot(input, output, session, state)
}
\description{
Fitplot creates a Population, Fit and Recommendation plot
}
