% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_fitPlot.R
\name{fitPlotUI}
\alias{fitPlotUI}
\title{This UI element displays several plots
that show the current state of the fit.}
\usage{
fitPlotUI(id, height = "auto")
}
\arguments{
\item{id}{output ID}

\item{height}{height of the plotly plots
Use "80vh" to cover 80\% of the screen space, or use a pixel amount to have a fixed size plot.}
}
\description{
It is displayed in a tabsetPanel
}
