% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{fromConfig}
\alias{fromConfig}
\title{Get the value corresponding to the given key from the configuration.}
\usage{
fromConfig(key)
}
\arguments{
\item{key}{a key}
}
\value{
the value corresponding to the key
}
\description{
Get the value corresponding to the given key from the configuration.
}
