% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.R
\name{getAllPatients}
\alias{getAllPatients}
\title{Get all patients from the database.}
\usage{
getAllPatients()
}
\value{
all the patients in a dataframe
}
\description{
Get all patients from the database.
}
