% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeUtilities.R
\name{getAppTimeZone}
\alias{getAppTimeZone}
\title{Get the application timezone.}
\usage{
getAppTimeZone()
}
\value{
by default, it returns the system timezone
}
\description{
Get the application timezone.
}
