% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariateUtilities.R
\name{getCovariateNames}
\alias{getCovariateNames}
\title{Get all covariate names of a model.
By default, covariate names from the model are returned.
For MPC models, only the 'true' covariates are returned (e.g. WT, not TVCl).}
\usage{
getCovariateNames(model)
}
\arguments{
\item{model}{tdmore model}
}
\value{
a character vector with the covariates
}
\description{
Get all covariate names of a model.
By default, covariate names from the model are returned.
For MPC models, only the 'true' covariates are returned (e.g. WT, not TVCl).
}
