% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeUtilities.R
\name{getDatetimeFormat}
\alias{getDatetimeFormat}
\title{Get the datetime format for this project.}
\usage{
getDatetimeFormat()
}
\value{
the datetime format
}
\description{
Get the datetime format for this project.
}
