% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeUtilities.R
\name{getDBDatetimeFormat}
\alias{getDBDatetimeFormat}
\title{Get the DB datetime format for this project.}
\usage{
getDBDatetimeFormat()
}
\value{
the datetime format
}
\description{
Get the DB datetime format for this project.
}
