% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{getDefaultModel}
\alias{getDefaultModel}
\title{Get default model.
If the given model is not a mixture model, this model is returned.
If the given model is a mixture_model, the most likely model is returned.}
\usage{
getDefaultModel(model)
}
\arguments{
\item{model}{tdmore model or tdmore_mixture model}
}
\value{
a tdmore model
}
\description{
Get default model.
If the given model is not a mixture model, this model is returned.
If the given model is a mixture_model, the most likely model is returned.
}
