% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtilities.R
\name{getDoseColumnLabel}
\alias{getDoseColumnLabel}
\title{Get dose column label.}
\usage{
getDoseColumnLabel(model, breakLine = T)
}
\arguments{
\item{model}{tdmore model}

\item{breakLine}{breakline or just space before the unit}
}
\value{
a label
}
\description{
Get dose column label.
}
