% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtilities.R
\name{getDosingInterval}
\alias{getDosingInterval}
\title{Retrieve the dosing interval from the tdmore metadata.
If no dose metadata is defined in the model, 24 hours is returned by default.}
\usage{
getDosingInterval(model)
}
\arguments{
\item{model}{tdmore model}
}
\value{
the dosing interval
}
\description{
Retrieve the dosing interval from the tdmore metadata.
If no dose metadata is defined in the model, 24 hours is returned by default.
}
