% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtilities.R
\name{getFormulationList}
\alias{getFormulationList}
\title{Get the list of formulations for the given model.}
\usage{
getFormulationList(model)
}
\arguments{
\item{model}{tdmore model}
}
\value{
a label
}
\description{
Get the list of formulations for the given model.
}
