% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtilities.R
\name{getMeasureColumnLabel}
\alias{getMeasureColumnLabel}
\title{Get measure column label.}
\usage{
getMeasureColumnLabel(model, breakLine = T)
}
\arguments{
\item{model}{tdmore model}

\item{breakLine}{breakline or just space before the unit}
}
\value{
a label
}
\description{
Get measure column label.
}
