% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newPatientDialog.R
\name{getModelList}
\alias{getModelList}
\title{Get the list of available models in this package.}
\usage{
getModelList()
}
\value{
a list of all model names
}
\description{
Get the list of available models in this package.
}
