% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtilities.R
\name{getModelOutput}
\alias{getModelOutput}
\title{Get model output variable name.}
\usage{
getModelOutput(model)
}
\arguments{
\item{model}{tdmore model}
}
\value{
the model output variable name, like 'CONC'
}
\description{
Get model output variable name.
}
