% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiMetadata.R
\name{getObservedVariables}
\alias{getObservedVariables}
\title{Get observed variables from tdmore model.}
\usage{
getObservedVariables(model)
}
\arguments{
\item{model}{tdmore model}
}
\value{
a list of observed variables
}
\description{
Get observed variables from tdmore model.
}
