% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence_layer.R
\name{getPatient}
\alias{getPatient}
\title{Get a patient from the database.}
\usage{
getPatient(id = NULL, firstname = NULL, lastname = NULL)
}
\arguments{
\item{id}{by ID}

\item{firstname}{by firstname}

\item{lastname}{by lastname}
}
\value{
the requested patient
}
\description{
Get a patient from the database.
}
