% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictionTab.R
\name{getPredictionTabPanel}
\alias{getPredictionTabPanel}
\title{Get the prediction tab panel.}
\usage{
getPredictionTabPanel()
}
\value{
a panel
}
\description{
Get the prediction tab panel.
}
