% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtilities.R
\name{getRecommendedDoseColumnLabel}
\alias{getRecommendedDoseColumnLabel}
\title{Get recommended dose column label.}
\usage{
getRecommendedDoseColumnLabel(model, breakLine = T)
}
\arguments{
\item{model}{tdmore model}

\item{breakLine}{breakline or just space before the unit}
}
\value{
a label
}
\description{
Get recommended dose column label.
}
