% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictionCore.R
\name{getRoundFunction}
\alias{getRoundFunction}
\title{Retrieve the rounding function from the tdmore metadata according to the specified formulation.
If no formulation matches, no rounding function is used.
If no formulation at all in metadata, no rounding function is used.}
\usage{
getRoundFunction(model, formulation = NULL)
}
\arguments{
\item{model}{tdmore model}

\item{formulation}{formulation we are looking for, character, can be NULL}
}
\value{
a round function, nevel NULL
}
\description{
Retrieve the rounding function from the tdmore metadata according to the specified formulation.
If no formulation matches, no rounding function is used.
If no formulation at all in metadata, no rounding function is used.
}
