% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeTable.R
\name{getTableBorder}
\alias{getTableBorder}
\title{Utility function to know where the horizontal line corresponding to the now date should be}
\usage{
getTableBorder(df, now)
}
\arguments{
\item{df}{data.frame with at least column 'time'}

\item{now}{now date}
}
\value{
empty list if no border required, or a list() according to https://handsontable.com/docs/7.3.0/Options.html#customBorders

Please note that handsontable.js has a bug in the handling of customBorders
See https://github.com/jrowen/rhandsontable/issues/336
}
\description{
Utility function to know where the horizontal line corresponding to the now date should be
}
