% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableUtilities.R
\name{getTableBorderIndex}
\alias{getTableBorderIndex}
\title{Utility function to know where the horizontal line corresponding to the now date should be.}
\usage{
getTableBorderIndex(data, now)
}
\arguments{
\item{data}{a data frame that has a date and time column}

\item{now}{now date}
}
\value{
index of this horizontal line or integer(0) if not pertinent
}
\description{
Utility function to know where the horizontal line corresponding to the now date should be.
}
