% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeUtilities.R
\name{getTimeFormat}
\alias{getTimeFormat}
\title{Get the time format for this project.}
\usage{
getTimeFormat()
}
\value{
the time format
}
\description{
Get the time format for this project.
}
