% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{getWinnerFit}
\alias{getWinnerFit}
\title{Get winner fit.
If the given fit is not a tdmorefit mixture, this fit is returned.
If the given fit is a tdmorefit mixture, the winner fit is returned.}
\usage{
getWinnerFit(fit)
}
\arguments{
\item{fit}{tdmore fit object}
}
\value{
a tdmore model
}
\description{
Get winner fit.
If the given fit is not a tdmorefit mixture, this fit is returned.
If the given fit is a tdmorefit mixture, the winner fit is returned.
}
