% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtilities.R
\name{getYAxisLabel}
\alias{getYAxisLabel}
\title{Get a nice Y-axis label.}
\usage{
getYAxisLabel(model)
}
\arguments{
\item{model}{tdmore model}
}
\value{
a label
}
\description{
Get a nice Y-axis label.
}
