% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newPatientDialog.R
\name{hackSelectInput}
\alias{hackSelectInput}
\title{Hack select input (update it with fake values when dialog is closed).
Without this workaround, covariate inputs do not appear when 'new patient' is clicked for the second time.}
\usage{
hackSelectInput(session)
}
\arguments{
\item{session}{shiny session}
}
\description{
Hack select input (update it with fake values when dialog is closed).
Without this workaround, covariate inputs do not appear when 'new patient' is clicked for the second time.
}
