% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeTable.R
\name{hot_to_r_datetime}
\alias{hot_to_r_datetime}
\title{Conversion function to convert an rhandsontable created with \link{timeTable} back
to a data.frame with a single `time` column.}
\usage{
hot_to_r_datetime(x)
}
\arguments{
\item{x}{rhandsontable output object}
}
\value{
data-frame with a single `time` column
}
\description{
Conversion function to convert an rhandsontable created with \link{timeTable} back
to a data.frame with a single `time` column.
}
