% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeUtilities.R
\name{hoursList}
\alias{hoursList}
\title{Generate the list of hours that can be picked in the 'hours' combobox.}
\usage{
hoursList()
}
\value{
a list of all the hours
}
\description{
Generate the list of hours that can be picked in the 'hours' combobox.
}
