% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientsTab.R
\name{initDataTable}
\alias{initDataTable}
\title{Collect data from the mongoDB.}
\usage{
initDataTable(ns, db)
}
\arguments{
\item{ns}{namespace}

\item{db}{database}
}
\description{
Collect data from the mongoDB.
}
