% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\docType{class}
\name{InMemoryDatabase}
\alias{InMemoryDatabase}
\title{InMemoryDatabase Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
InMemoryDatabase
}
\description{
InMemoryDatabase class description
}
\details{
In-memory database class
}
\keyword{datasets}
