% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientModel.R
\name{isPrivatePatient}
\alias{isPrivatePatient}
\title{Say if the given patient is private.}
\usage{
isPrivatePatient(patientModel)
}
\arguments{
\item{patientModel}{the patient model}
}
\value{
a logical value
}
\description{
Say if the given patient is private.
}
