% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientModel.R
\name{isReadOnlyPatient}
\alias{isReadOnlyPatient}
\title{Say if the given patient is read-only.}
\usage{
isReadOnlyPatient(patientModel)
}
\arguments{
\item{patientModel}{the patient model}
}
\value{
a logical value
}
\description{
Say if the given patient is read-only.
}
