% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\docType{class}
\name{JsonDatabase}
\alias{JsonDatabase}
\title{JsonDatabase Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
JsonDatabase
}
\description{
JsonDatabase class description
}
\details{
JSON database class
}
\keyword{datasets}
