% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientModel.R
\name{jsonToCovariateModel}
\alias{jsonToCovariateModel}
\title{Convert covariates (JSON) to covariates model.}
\usage{
jsonToCovariateModel(covariateJson)
}
\arguments{
\item{covariateJson}{covariates, JSON form}
}
\value{
tibble with date, time, covariate name and value  columns
}
\description{
Convert covariates (JSON) to covariates model.
}
