% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientModel.R
\name{jsonToDoseModel}
\alias{jsonToDoseModel}
\title{Convert doses (JSON) to dose model.}
\usage{
jsonToDoseModel(doseJson)
}
\arguments{
\item{doseJson}{doses, JSON form}
}
\value{
tibble with date, time and dose columns
}
\description{
Convert doses (JSON) to dose model.
}
