% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientModel.R
\name{jsonToMeasureModel}
\alias{jsonToMeasureModel}
\title{Convert measures (JSON) to dose model.}
\usage{
jsonToMeasureModel(measureJson)
}
\arguments{
\item{measureJson}{measures, JSON form}
}
\value{
tibble with date, time and dose columns
}
\description{
Convert measures (JSON) to dose model.
}
