% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientModel.R
\name{jsonToPatientModel}
\alias{jsonToPatientModel}
\title{Convert patient (JSON) to patient model}
\usage{
jsonToPatientModel(patientJson)
}
\arguments{
\item{patientJson}{the patient, JSON form}
}
\value{
the patient model
}
\description{
Convert patient (JSON) to patient model
}
