% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientModel.R
\name{measureModelToJson}
\alias{measureModelToJson}
\title{Convert measure model to JSON structure.}
\usage{
measureModelToJson(measureModel)
}
\arguments{
\item{measureModel}{the measure model}
}
\value{
data frame with date column (full date) and measure column
}
\description{
Convert measure model to JSON structure.
}
