% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataConversion.R
\name{mergeFormAndCov}
\alias{mergeFormAndCov}
\title{Join covariates and formulations (shinyTDMore -> TDMore).}
\usage{
mergeFormAndCov(covs, doses)
}
\arguments{
\item{covs}{shinyTDMore covariates}

\item{doses}{shinyTDMore doses}
}
\value{
TDMore covariates
}
\description{
Join covariates and formulations (shinyTDMore -> TDMore).
}
