% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtilities.R
\name{mergePlots}
\alias{mergePlots}
\title{Merge plots.}
\usage{
mergePlots(p1, p2, p3, output, source)
}
\arguments{
\item{p1}{first plot}

\item{p2}{second plot}

\item{p3}{third plot}

\item{output}{main tdmore output}
}
\description{
Merge plots.
}
