% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_modelTab.R
\name{modelTab}
\alias{modelTab}
\title{Model tab server.}
\usage{
modelTab(
  input,
  output,
  session,
  state,
  rmdFile = system.file("model.Rmd", package = "shinytdmore")
)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{state}{a reactiveValues object with the state of the application. See [convertDataToTdmore()] for more information.}

\item{rmdFile}{Rmarkdown file to create a description of the model}
}
\description{
Model tab server.
}
