% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTab.R
\name{modelTabUI}
\alias{modelTabUI}
\title{Model tab user interface.}
\usage{
modelTabUI(id)
}
\arguments{
\item{id}{namespace id}
}
\value{
a panel
}
\description{
Model tab user interface.
}
