% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_mongodb.R
\docType{class}
\name{MongoDatabase}
\alias{MongoDatabase}
\title{MongoDatabase Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MongoDatabase
}
\description{
MongoDatabase class description
}
\details{
Mongo database class
}
\keyword{datasets}
