% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataConversion.R
\name{nearEqual}
\alias{nearEqual}
\title{Near (in)equality function.
Different modes:
'ae' : almost equal
'gt' : greater than
'lt' : less than
'ne.gt' : greater than (near equality)
'ne.lt' : less than (near equality)}
\usage{
nearEqual(x, y, mode = "ae", tol = 1e-08)
}
\arguments{
\item{x}{first vector}

\item{y}{second vector}

\item{mode}{comparison mode}

\item{tol}{tolerance}
}
\value{
logical vector
}
\description{
Near (in)equality function.
Different modes:
'ae' : almost equal
'gt' : greater than
'lt' : less than
'ne.gt' : greater than (near equality)
'ne.lt' : less than (near equality)
}
