% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newPatientDialog.R
\name{newPatientDialogServer}
\alias{newPatientDialogServer}
\title{New patient dialog server.}
\usage{
newPatientDialogServer(input, output, session)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\description{
New patient dialog server.
}
