% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_now.R
\name{nowInput}
\alias{nowInput}
\alias{nowInputUI}
\title{Input element representing the current time}
\usage{
nowInputUI(id)

nowInput(input, output, session, state, debug = FALSE)
}
\arguments{
\item{id}{unique identifier}

\item{input}{input object}

\item{output}{output object}

\item{session}{session object}

\item{state}{reactiveValues object with value `now`. The server logic ensures a synchronization between `state$now` and the nowInput element.}

\item{debug}{whether to print debugging messages}
}
\description{
Input element representing the current time
}
