% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTab.R
\name{onModelTabExit}
\alias{onModelTabExit}
\title{Say if the user went out of the 'Model' tab.}
\usage{
onModelTabExit(onTabChanged)
}
\arguments{
\item{onTabChanged}{onTabChanged reactive values}

\item{input}{shiny input object}
}
\value{
a logical value
}
\description{
Say if the user went out of the 'Model' tab.
}
