% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveProject.R
\name{onPredictionTabExit}
\alias{onPredictionTabExit}
\title{Say if the user went out of the 'Prediction' tab.}
\usage{
onPredictionTabExit(onTabChanged)
}
\arguments{
\item{onTabChanged}{onTabChanged reactive values}

\item{input}{shiny input object}
}
\value{
a logical value
}
\description{
Say if the user went out of the 'Prediction' tab.
}
