% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeUtilities.R
\name{pad}
\alias{pad}
\title{Pad integer with a zero if needed.}
\usage{
pad(integer)
}
\arguments{
\item{integer}{the integer to be padded}
}
\value{
a string
}
\description{
Pad integer with a zero if needed.
}
