% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newPatientDialog.R
\name{patientFormModalDialog}
\alias{patientFormModalDialog}
\title{Create a patient form in a modal dialog.}
\usage{
patientFormModalDialog(failed = FALSE)
}
\arguments{
\item{failed}{logical value, can be true if there was an error in the form}
}
\description{
Create a patient form in a modal dialog.
}
