% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientModel.R
\name{patientModelToJson}
\alias{patientModelToJson}
\title{Convert patient model to JSON structure.}
\usage{
patientModelToJson(patientModel)
}
\arguments{
\item{patientModel}{the patient model}
}
\value{
the patient, JSON form
}
\description{
Convert patient model to JSON structure.
}
