% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientsTab.R
\name{patientsTab}
\alias{patientsTab}
\title{Patients tab server.}
\usage{
patientsTab(input, output, session, parentSession, val, onNewPatientAdded,
  db)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{parentSession}{shiny parent session}

\item{val}{main reactive container}

\item{onNewPatientAdded}{reactive value}

\item{db}{database}
}
\description{
Patients tab server.
}
