% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientsTab.R
\name{patientsTabServer}
\alias{patientsTabServer}
\title{Patients tab server.}
\usage{
patientsTabServer(input, output, session, val)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{val}{main reactive container}
}
\description{
Patients tab server.
}
