% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientsTab.R
\name{patientsTabUI}
\alias{patientsTabUI}
\title{Patients tab user interface.}
\usage{
patientsTabUI(id)
}
\arguments{
\item{id}{namespace id}
}
\value{
a panel
}
\description{
Patients tab user interface.
}
