% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeUtilities.R
\name{POSIXToDate}
\alias{POSIXToDate}
\title{Format a POSIX datetime to a 'Date' object.
Note that the returned string will depend on the application timezone.}
\usage{
POSIXToDate(posixDatetime)
}
\arguments{
\item{posixDatetime}{POSIX datetime}
}
\value{
a 'Date' object
}
\description{
Format a POSIX datetime to a 'Date' object.
Note that the returned string will depend on the application timezone.
}
