% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeUtilities.R
\name{POSIXToHours}
\alias{POSIXToHours}
\title{Convert POSIX date to hours (numeric).}
\usage{
POSIXToHours(posixDate)
}
\arguments{
\item{posixDate}{a posix date}
}
\value{
the converted dates, in hours
}
\description{
Convert POSIX date to hours (numeric).
}
