% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.R
\name{posixToString}
\alias{posixToString}
\title{Convert a posix date to a string.}
\usage{
posixToString(posixDate)
}
\arguments{
\item{posixDate}{a posix date}
}
\value{
a well formated string corresponding to the date, time zone included
}
\description{
Convert a posix date to a string.
}
