% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeUtilities.R
\name{POSIXToTime}
\alias{POSIXToTime}
\title{Format a POSIX datetime to a time string.
Note that the returned string will depend on the application timezone.}
\usage{
POSIXToTime(posixDatetime)
}
\arguments{
\item{posixDatetime}{POSIX datetime}
}
\value{
a time string formatted according to getTimeFormat()
}
\description{
Format a POSIX datetime to a time string.
Note that the returned string will depend on the application timezone.
}
