% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictionTab.R
\name{predictionTab}
\alias{predictionTab}
\title{Prediction tab server.}
\usage{
predictionTab(input, output, session, state)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{state}{program state, containing the fields `now`, `model`, `regimen`, `observed`}
}
\description{
Prediction tab server.
}
