% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictionTab.R
\name{predictionTabServer}
\alias{predictionTabServer}
\title{Prediction tab server.}
\usage{
predictionTabServer(input, output, session, val)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{val}{main reactive container}
}
\description{
Prediction tab server.
}
