% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtilities.R
\name{prepareParametersPlot}
\alias{prepareParametersPlot}
\title{Prepare the parameters plot.}
\usage{
prepareParametersPlot(populationPredict, individualPredict, parameters)
}
\arguments{
\item{populationPredict}{dataframe containing the individual prediction data}

\item{individualPredict}{dataframe containing the population prediction data}

\item{parameters}{parameters}
}
\description{
Prepare the parameters plot.
}
