% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtilities.R
\name{prepareRecommendationPlots}
\alias{prepareRecommendationPlots}
\title{Prepare recommendation plots (to be refactored).
This includes the recommendation plot and the timeline plot.}
\usage{
prepareRecommendationPlots(
  populationDf,
  individualDf,
  recommendationDf,
  observed,
  target,
  model,
  now,
  regimen,
  recommendedRegimen
)
}
\arguments{
\item{target}{numeric vector of size 2, min and max value}

\item{model}{tdmore model}

\item{now}{now date}

\item{doses}{doses}

\item{obs}{observations}

\item{covs}{covariates}

\item{recommendationData}{recommendation data (contains namely ipred, ipredNew and the recommendation)}
}
\value{
a list of two plots
}
\description{
Prepare recommendation plots (to be refactored).
This includes the recommendation plot and the timeline plot.
}
