% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtilities.R
\name{prepareRecommendedTimelinePlot}
\alias{prepareRecommendedTimelinePlot}
\title{Prepare the timeline with both original and recommended doses.}
\usage{
prepareRecommendedTimelinePlot(regimen, newRegimen, model, now)
}
\arguments{
\item{model}{tdmore model}

\item{now}{now date, currently not used here}

\item{originalDoses}{dataframe containing the original entered doses}

\item{recommendedDoses}{dataframe containing the recommended doses calculated by the computer}

\item{xlim}{ggplot xlim argument}
}
\description{
Prepare the timeline with both original and recommended doses.
}
