% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictionTab.R
\name{previousNextLogic}
\alias{previousNextLogic}
\title{Previous/Next button logic.}
\usage{
previousNextLogic(input, output, session, val, outputReact, ns)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{val}{main reactive container}

\item{outputReact}{reactive value for the plot type}

\item{ns}{namespace}
}
\description{
Previous/Next button logic.
}
