% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtilities.R
\name{recommendationColor}
\alias{recommendationColor}
\title{Recommendation color (green).}
\usage{
recommendationColor()
}
\value{
a color string
}
\description{
Recommendation color (green).
}
