% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_hotTable.R
\name{recommendationTable}
\alias{recommendationTable}
\title{The recommendationTable uses the state$regimen and state$recommendation tables}
\usage{
recommendationTable(input, output, session, state)
}
\description{
The recommendationTable uses the state$regimen and state$recommendation tables
}
