% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTab.R
\name{rejectModel}
\alias{rejectModel}
\title{Reject new selected model.}
\usage{
rejectModel(val, session)
}
\arguments{
\item{val}{main reactive container}

\item{session}{shiny session}
}
\description{
Reject new selected model.
}
