% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.R
\name{removePatient}
\alias{removePatient}
\title{Remove a patient from the database.}
\usage{
removePatient(id)
}
\arguments{
\item{id}{the patient ID}
}
\description{
Remove a patient from the database.
}
