% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientsTab.R
\name{removePatientConfirmationDialog}
\alias{removePatientConfirmationDialog}
\title{Remove patient confirmation dialog.}
\usage{
removePatientConfirmationDialog(patientName, ns)
}
\arguments{
\item{patientName}{the patient name}

\item{ns}{namespace}
}
\value{
a modal dialog
}
\description{
Remove patient confirmation dialog.
}
