% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientsTab.R
\name{renderPatientTable}
\alias{renderPatientTable}
\title{Render the patients table.}
\usage{
renderPatientTable(input)
}
\arguments{
\item{input}{the shiny input object}
}
\description{
Render the patients table.
}
