% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportsTab.R
\name{reportsTab}
\alias{reportsTab}
\title{Reports tab server.}
\usage{
reportsTab(input, output, session, val)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{val}{main reactive container}
}
\description{
Reports tab server.
}
