% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportsTab.R
\name{reportsTabUI}
\alias{reportsTabUI}
\title{Reports tab user interface.}
\usage{
reportsTabUI(id)
}
\arguments{
\item{id}{namespace id}
}
\value{
a panel
}
\description{
Reports tab user interface.
}
