% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newPatientDialog.R
\name{saveData}
\alias{saveData}
\title{Save data of both user and covariate forms.}
\usage{
saveData(userData, modelName, covariateData, db)
}
\arguments{
\item{userData}{data from user from, named vector}

\item{modelName}{model name}

\item{covariateData}{data from covariate form, named vector}
}
\description{
Save data of both user and covariate forms.
}
