% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveProject.R
\name{saveProject}
\alias{saveProject}
\title{Save project into DB (measures and doses from shiny are saved).}
\usage{
saveProject(val)
}
\arguments{
\item{val}{main reactive container}
}
\description{
Save project into DB (measures and doses from shiny are saved).
}
