% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveProject.R
\name{saveProjectServer}
\alias{saveProjectServer}
\title{Save project server}
\usage{
saveProjectServer(input, val)
}
\arguments{
\item{input}{shiny input object}

\item{val}{main reactive container}
}
\description{
Save project server
}
