% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveProject.R
\name{saveProjectToDB}
\alias{saveProjectToDB}
\title{Save project into DB (measures and doses from shiny are saved).}
\usage{
saveProjectToDB(val, db)
}
\arguments{
\item{val}{main reactive container}
}
\description{
Save project into DB (measures and doses from shiny are saved).
}
