% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTdmore.R
\name{selectPatient}
\alias{selectPatient}
\title{Select patient (last in DB or from URL).}
\usage{
selectPatient(session, val, db)
}
\arguments{
\item{session}{shiny session}

\item{val}{main reactive container}

\item{db}{database}
}
\description{
Select patient (last in DB or from URL).
}
