% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTdmore.R
\name{selectPatientFromURL}
\alias{selectPatientFromURL}
\title{Select patient from URL logic.}
\usage{
selectPatientFromURL(session, val)
}
\arguments{
\item{session}{shiny session}

\item{val}{main reactive container}
}
\description{
Select patient from URL logic.
}
