% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_form.R
\name{server}
\alias{server}
\title{Server inputs/outputs.}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\description{
Server inputs/outputs.
}
