% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientsTab.R
\name{setPatient}
\alias{setPatient}
\title{Set patient to shiny app.}
\usage{
setPatient(patient, val)
}
\arguments{
\item{patient}{the patient to be set}

\item{val}{main reactive container}
}
\description{
Set patient to shiny app.
}
