% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyUtilities.R
\name{shinyInput}
\alias{shinyInput}
\title{Allow to provide HTML/javascript code in shiny app for custom inputs.
Example if FUN is actionButton:
<button id=\"removeButton_1\" type=\"button\" class=\"btn btn-default action-button\" onclick=\"Shiny.onInputChange(&quot;removePatientButton&quot;,  this.id)\">Remove</button>".}
\usage{
shinyInput(FUN, id, label, ...)
}
\arguments{
\item{FUN}{a compatible shiny function (like actionButton or actionLink)}

\item{id}{id used to identify the inputs, character}

\item{label}{input label, character vector, determines output size}

\item{...}{additional arguments}
}
\value{
a character vector (length depends on label size)
}
\description{
Allow to provide HTML/javascript code in shiny app for custom inputs.
Example if FUN is actionButton:
<button id=\"removeButton_1\" type=\"button\" class=\"btn btn-default action-button\" onclick=\"Shiny.onInputChange(&quot;removePatientButton&quot;,  this.id)\">Remove</button>".
}
