% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zPackage.R
\name{shinytdmore-data}
\alias{shinytdmore-data}
\title{Data definition}
\arguments{
\item{model}{a tdmore model with (ideally) metadata}

\item{regimen}{a data.frame with the following columns:
\itemize{
\item \code{time} a POSIXct time
\item \code{dose} a numeric with the dose amount
\item \code{formulation} a character vector that corresponds to formulations defined in the metadata of the model
\item \code{fix} a boolean vector describing whether the given regimen can be modified in dose recommendation
}}

\item{observed}{a data.frame with the following columns:
\itemize{
\item \code{time} a POSIXct time
\item \code{dv} the observed data
\item \code{use} whether the observation should be used in the estimation
}}

\item{covariates}{a data.frame with the following columns:
\itemize{
\item \code{time} a POSIXct time
\item \code{...} columns corresponding to covariates required in the tdmore model. If columns are missing, the method returns an error.
}}

\item{now}{a POSIXct time representing the current time}

\item{target}{a named list with min/max values to represent the target concentration}
}
\description{
Data definition
}
