% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zPackage.R
\name{shinytdmore-module}
\alias{shinytdmore-module}
\title{Basic definition for module server components}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{state}{a reactiveValues object with the state of the application. See [convertDataToTdmore()] for more information.}

\item{cr}{list of calculation-related reactives, see [calculationReactives()]}
}
\description{
Basic definition for module server components
}
