% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtilities.R
\name{shinytdmore-plot}
\alias{shinytdmore-plot}
\title{Default values for plot parameters}
\arguments{
\item{individualPredict}{dataframe containing the population prediction data}

\item{populationPredict}{dataframe containing the individual prediction data}

\item{recommendationPredict}{recommendation data (contains namely ipred, ipredNew and the recommendation)}

\item{parameters}{parameters}

\item{recommendedRegimen}{regimen with the AMT column changed by the recommended doses}
}
\description{
Default values for plot parameters
}
