% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTdmore.R
\name{shinyTdmore}
\alias{shinyTdmore}
\title{Server component of Shiny TDMore.}
\usage{
shinyTdmore(input, output, session, conf, db)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{conf}{modules configuration
This is a named list of extra shiny modules to load. Each module will receive the arguments 
`val` (a reactiveValues() set) and `db` (the database)}

\item{db}{database, uses in-memory database if missing}
}
\value{
server component
}
\description{
The server expects the following UI elements:
- input$tabs
-
}
