% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appMain.R
\name{shinyTdmoreServer}
\alias{shinyTdmoreServer}
\title{Get server component of Shiny TDMore.}
\usage{
shinyTdmoreServer(input, output, session)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\value{
server component
}
\description{
Get server component of Shiny TDMore.
}
