% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTdmore.R
\name{shinyTdmoreUI}
\alias{shinyTdmoreUI}
\title{Get UI component of Shiny TDMore.}
\usage{
shinyTdmoreUI(title, ...)
}
\arguments{
\item{title}{shiny app title}

\item{...}{UI functions}
}
\value{
UI component
}
\description{
Get UI component of Shiny TDMore.
}
