% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{ShinyToDplyrProgressFacade}
\alias{ShinyToDplyrProgressFacade}
\title{A facade object}
\description{
A facade object

A facade object
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ShinyToDplyrProgressFacade$new()}}
\item \href{#method-tick}{\code{ShinyToDplyrProgressFacade$tick()}}
\item \href{#method-stop}{\code{ShinyToDplyrProgressFacade$stop()}}
\item \href{#method-print}{\code{ShinyToDplyrProgressFacade$print()}}
\item \href{#method-clone}{\code{ShinyToDplyrProgressFacade$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyToDplyrProgressFacade$new(
  n = 1,
  min_time = 0,
  proxy = NULL,
  amount = 1,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tick"></a>}}
\subsection{Method \code{tick()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyToDplyrProgressFacade$tick()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stop"></a>}}
\subsection{Method \code{stop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyToDplyrProgressFacade$stop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyToDplyrProgressFacade$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyToDplyrProgressFacade$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
