% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeTable.R
\name{singleReactive}
\alias{singleReactive}
\title{Create a function that behaves as a reactiveVal(), but is backed by a reactiveValues() object}
\usage{
singleReactive(state, key, default = NULL, from = identity, to = identity)
}
\arguments{
\item{state}{the reactiveValues object}

\item{key}{character key}

\item{default}{default value to return if `state$key` is null}

\item{from}{function to convert value from state into output}

\item{to}{function to convert from input into state}
}
\description{
Create a function that behaves as a reactiveVal(), but is backed by a reactiveValues() object
}
