% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeUtilities.R
\name{stringToPOSIX}
\alias{stringToPOSIX}
\title{Convert a string date to a POSIX datetime.}
\usage{
stringToPOSIX(stringDatetime)
}
\arguments{
\item{stringDatetime}{a string datetime that includes the zone, formatted this way: '2018-12-08 09:00:00 +0100'}
}
\value{
a POSIX datetime
}
\description{
Convert a string date to a POSIX datetime.
}
