% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_synchronizedHot.R
\name{synchronizedHot}
\alias{synchronizedHot}
\alias{synchronizedHotUi}
\title{An rhandsontable that is synchronized to a reactiveValues element}
\usage{
synchronizedHotUi(id, ...)

synchronizedHot(
  input,
  output,
  session,
  stateDf,
  expr,
  hot_to_r = rhandsontable::hot_to_r,
  debug = FALSE
)
}
\arguments{
\item{id}{unique identifier}

\item{...}{passed to [rhandsontable::rHandsontableOutput()]}

\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{stateDf}{a reactiveVal data.frame used to write the state}

\item{expr}{expression used to create a rhandsontable. Make sure to isolate() calls to the state$df object. The observers
in this module will make sure the handsontable is recreated if a relevant change to state$df occurs.}

\item{hot_to_r}{a function to convert the input$table object into a data.frame of the same format as stateDf}

\item{debug}{TRUE to enable printing debug messages}
}
\description{
An rhandsontable that is synchronized to a reactiveValues element
}
