% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_hotTable.R
\name{tableUI}
\alias{tableUI}
\title{Create a user interface to represent a table, with a "table" UI element and an "Add" button}
\usage{
tableUI(id, ...)
}
\arguments{
\item{id}{id for the encapsulating div}

\item{...}{extra arguments passed to actionButton}
}
\description{
Create a user interface to represent a table, with a "table" UI element and an "Add" button
}
