% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{tacrolimuskidney}
\alias{tacrolimuskidney}
\title{Tacrolimus kidney model.}
\format{RxODE model}
\usage{
tacrolimuskidney
}
\description{
Tacrolimus kidney model.
}
\keyword{datasets}
