% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_target.R
\name{targetInput}
\alias{targetInput}
\alias{targetInputUI}
\title{Input element representing the target upper limit and lower limits}
\usage{
targetInputUI(id)

targetInput(input, output, session, state)
}
\arguments{
\item{id}{unique identifier}
}
\description{
Input element representing the target upper limit and lower limits
}
