% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictionTab.R
\name{targetLogic}
\alias{targetLogic}
\title{Target logic.}
\usage{
targetLogic(input, output, session, val, ns)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{val}{main reactive container}

\item{ns}{namespace}
}
\description{
Target logic.
}
