% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{testDBConfig}
\alias{testDBConfig}
\title{Get the DB configuration for tests.}
\usage{
testDBConfig()
}
\value{
a shinytdmore_db_config object
}
\description{
Get the DB configuration for tests.
}
