% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeTable.R
\name{timeTable}
\alias{timeTable}
\title{Create an rhandsontable based on the given data-frame}
\usage{
timeTable(df, borders = list(), colHeaders = c())
}
\arguments{
\item{df}{data-frame with at least a 'time' column as the first column}

\item{borders}{a list() with custom borders. See getTableBorder() for more information.}

\item{colHeaders}{headers for the data-frame *excluding* the time column.}
}
\value{
an rhandsontable with columns Date, Time and any extra columns specified in colHeaders
}
\description{
Create an rhandsontable based on the given data-frame
}
\seealso{
hot_to_r_datetime
}
