% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{toConfig}
\alias{toConfig}
\title{Add key/value pair to configuration.}
\usage{
toConfig(key, value)
}
\arguments{
\item{key}{a key}

\item{value}{any object}
}
\description{
Add key/value pair to configuration.
}
