% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo.R
\name{toJSONPatient}
\alias{toJSONPatient}
\title{Convert patient to JSON format.}
\usage{
toJSONPatient(patient)
}
\arguments{
\item{patient}{the patient}
}
\value{
a JSON string for MongoDB
}
\description{
Convert patient to JSON format.
}
