% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientModel.R
\name{toPrivatePatient}
\alias{toPrivatePatient}
\title{Classify a patient as private.
These patients will not appear in the patients tab.}
\usage{
toPrivatePatient(patientModel)
}
\arguments{
\item{patientModel}{the patient model}
}
\value{
the patient model
}
\description{
Classify a patient as private.
These patients will not appear in the patients tab.
}
