% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientModel.R
\name{toReadOnlyPatient}
\alias{toReadOnlyPatient}
\title{Make a patient read-only.}
\usage{
toReadOnlyPatient(patientModel)
}
\arguments{
\item{patientModel}{the patient model}
}
\value{
the patient model
}
\description{
Make a patient read-only.
}
