% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientForm.R
\docType{data}
\name{ui}
\alias{ui}
\title{Define the user interface components}
\format{An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 3.}
\usage{
ui
}
\description{
Define the user interface components
}
\keyword{datasets}
