% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newPatientDialog.R
\name{updateCovariates}
\alias{updateCovariates}
\title{Update covariates in patient model (convert numeric vector to tibble).}
\usage{
updateCovariates(patient, covs, date, time)
}
\arguments{
\item{patient}{patient model}

\item{covs}{numeric vector (or null if no covariates)}

\item{date}{date}

\item{time}{time (character)}
}
\value{
an updated patient model
}
\description{
Update covariates in patient model (convert numeric vector to tibble).
}
