% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xEditableInput.R
\name{updateEditableInput}
\alias{updateEditableInput}
\title{Update an editable input.}
\usage{
updateEditableInput(session, inputId, value = NULL)
}
\arguments{
\item{session}{inputId}

\item{inputId}{type}

\item{value}{value}

\item{options}{options}
}
\description{
Update an editable input.
}
