% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTab.R
\name{updateModelInSelectInput}
\alias{updateModelInSelectInput}
\title{Update the model in the select input component.}
\usage{
updateModelInSelectInput(selectedModel, session)
}
\arguments{
\item{selectedModel}{the new selected model}

\item{session}{shiny session}
}
\description{
Update the model in the select input component.
}
