% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientModel.R
\name{updateNowDate}
\alias{updateNowDate}
\title{Update the date of now.}
\usage{
updateNowDate(patient, now)
}
\arguments{
\item{patient}{the given patient}

\item{now}{the date of now, POSIXlt time}
}
\value{
the patient
}
\description{
Update the date of now.
}
