% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence_layer.R
\name{updatePatient}
\alias{updatePatient}
\title{Update a patient.}
\usage{
updatePatient(id, patient)
}
\arguments{
\item{id}{the ID to be updated}

\item{patient}{the updated patient}
}
\value{
the patient ID
}
\description{
Update a patient.
}
