% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientModel.R
\name{updatePatientCovariates}
\alias{updatePatientCovariates}
\title{Update patient covariates.}
\usage{
updatePatientCovariates(patient, covariates)
}
\arguments{
\item{patient}{the given patient}

\item{covariates}{the covariates, data frame}
}
\value{
the patient
}
\description{
Update patient covariates.
}
