% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientModel.R
\name{updatePatientDoses}
\alias{updatePatientDoses}
\title{Update patient doses.}
\usage{
updatePatientDoses(patient, doses)
}
\arguments{
\item{patient}{the given patient}

\item{doses}{the doses, data frame}
}
\value{
the patient
}
\description{
Update patient doses.
}
