% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.R
\name{updatePatientMeasures}
\alias{updatePatientMeasures}
\title{Update patient measures.}
\usage{
updatePatientMeasures(patient, measures)
}
\arguments{
\item{patient}{the given patient}

\item{measures}{the measures, data frame}
}
\value{
the patient
}
\description{
Update patient measures.
}
