% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.R
\name{updatePatientModel}
\alias{updatePatientModel}
\title{Update patient model and covariates}
\usage{
updatePatientModel(patient, modelName, covariates = c())
}
\arguments{
\item{patient}{the given patient}

\item{modelName}{model name}

\item{covariates}{named numeric with the covariates}
}
\value{
the patient
}
\description{
Update patient model and covariates
}
