% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_aboutTab.R
\name{aboutTabUI}
\alias{aboutTabUI}
\title{This tab shows general information out the application.}
\usage{
aboutTabUI(id, htmlFile)
}
\arguments{
\item{id}{namespace id}

\item{htmlFile}{the html file to use for the contents of the about tab. If missing, we use the default about.html from the package directory.}
}
\value{
a tabPanel with HTML content
}
\description{
This tab shows general information out the application.
}
