% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableUtilities.R
\name{autoSortByDate}
\alias{autoSortByDate}
\title{Auto-sort table by date and convert to tibble.}
\usage{
autoSortByDate(data)
}
\arguments{
\item{data}{a data frame that has a date and time column}
}
\value{
a tibble, ordered by date
}
\description{
Auto-sort table by date and convert to tibble.
}
