% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictionTab.R
\name{autoSortByDate}
\alias{autoSortByDate}
\title{Auto-sort doses table by date.}
\usage{
autoSortByDate(data)
}
\arguments{
\item{doses}{the given 'doses' data frame}
}
\value{
the same dataframe, ordered by date
}
\description{
Auto-sort doses table by date.
}
