% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{calculation}
\alias{calculation}
\title{This module performs all of the required calculation for the application.}
\usage{
calculation(state, millis = 500, .mc.maxpts = 300)
}
\arguments{
\item{state}{reactiveValues object with full state}

\item{millis}{time delay for reacting to changes in the input}

\item{.mc.maxpts}{number of points in monte carlo}
}
\description{
Call it in your application using `calculation(state)`.
}
\details{
Shiny progress calls are used to report which calculation is being done. However, the progress bar does not advance.
It is very difficult to create an encompassing progress bar throughout all steps, because some of these steps are shortcutted depending 
on the changes done. As an example, changes in future doses will not change the fitted parameters.
}
\section{Output}{

The module sets the following state variables:
* `populationPredict` a data-frame with prediction data on the population level
* `fit` a tdmorefit object
* `individualPredict` a data-frame with prediction data for the individual
* `recommendation` a data.frame in the same form as `regimen`, but where the AMT column was adapted

If any of these values could not be calculated, the outputs are set to the error that occurred during calculation. This may be
a simple R error [simpleError()] or a more verbose shiny validation error (e.g. [shiny::validate()] or [shiny::req()])
}

