% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{calculation}
\alias{calculation}
\alias{reactiveFit}
\alias{reactiveRecommendation}
\alias{calculationReactives}
\alias{reactivePredict}
\title{Reactives that can be used to support tdmore calculations}
\usage{
reactiveFit(
  state,
  population = FALSE,
  millis = 2000,
  label = if (population) "PopulationFit" else "IndividualFit",
  estimate = tdmore::estimate
)

reactiveRecommendation(state, fit, millis = 2000)

calculationReactives(
  state,
  mc.maxpts = 100,
  fitMillis = 2000,
  predictMillis = 2000,
  recommendationMillis = 500,
  estimate = tdmore::estimate
)

reactivePredict(
  state,
  fit = NULL,
  recommendation = NULL,
  mc.maxpts = 100,
  millis = 2000
)
}
\arguments{
\item{state}{reactiveValues() or list-like object with input data, see \link{shinytdmore-data} for more information}

\item{population}{TRUE to calculate a population fit, FALSE to calculate an individual fit}

\item{millis}{time for debounce, see [shiny::debounce()]}

\item{label}{label for all reactives in this system}

\item{estimate}{estimation function, uses `tdmore::estimate` by default}

\item{fit}{reactive of type [reactiveFit()] with the current fit}

\item{mc.maxpts}{number of monte carlo samples to take, see [tdmore::predict.tdmorefit()]}

\item{fitMillis}{time to wait until calculating new fit}

\item{predictMillis}{time to wait until calculating a new prediction}

\item{recommendationMillis}{time to wait until calculating a new recommendation}

\item{recommendation}{recommendation reactive of type [reactiveRecommendation()]}
}
\description{
Reactives that can be used to support tdmore calculations
}
