% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtilities.R
\name{convertDataToTdmore}
\alias{convertDataToTdmore}
\title{Convert main data to TDMore domain.}
\usage{
convertDataToTdmore(doses, obs, covs, output, now, iov)
}
\arguments{
\item{doses}{doses}

\item{obs}{observations}

\item{covs}{covariates}

\item{output}{output name (e.g. 'CONC')}

\item{now}{now date, POSIXlt date}

\item{iov}{iov present in the model, logical value}
}
\value{
to be described
}
\description{
Convert main data to TDMore domain.
}
