% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataConversion.R
\name{convertDataToTdmore}
\alias{convertDataToTdmore}
\title{Convert shinyTDMore domain to TDMore domain}
\usage{
convertDataToTdmore(state)
}
\arguments{
\item{state}{reactiveValues() or list-like object with input data}
}
\value{
named list of tdmore-compatible values
}
\description{
Convert shinyTDMore domain to TDMore domain
}
\details{
Shinytdmore uses the following data objects in a reactiveValues() state:
\itemize{
\item \code{state$model} a tdmore model with (ideally) metadata
\item \code{state$regimen} a data.frame with the following columns:
\itemize{
\item \code{time} a POSIXct time
\item \code{dose} a numeric with the dose amount
\item \code{formulation} a character vector that corresponds to formulations defined in the metadata of the model
\item \code{fix} a boolean vector describing whether the given regimen can be modified in dose recommendation
}
\item \code{state$observed} a data.frame with the following columns:
\itemize{
\item \code{time} a POSIXct time
\item \code{dv} the observed data
\item \code{use} whether the observation should be used in the estimation
}
\item \code{state$covariates} a data.frame with the following columns:
\itemize{
\item \code{time} a POSIXct time
\item \code{...} columns corresponding to covariates required in the tdmore model. If columns are missing, the method returns an error.
}
\item \code{state$now} a POSIXct time representing the current time
}

This is converted into a list of arguments compatible with \link[tdmore:estimate]{tdmore::estimate}, \link[tdmore:predict.tdmore]{tdmore::predict.tdmore} and \link[tdmore:findDose]{tdmore::findDose}.
\itemize{
\item \code{model} the tdmore model
\item \code{regimen} all treatments in the right format
\item \code{observed} all observed values where use=TRUE, and with the \code{dv} column renamed to the default model output
\item \code{covariates} all covariates, with an added covariate \code{FORM} that contains the formulation
\item \code{t0} a POSIXct time corresponding to time \code{0}
\item \code{now} a numeric time, corresponding to the different between t0 and now
\item \code{doseRows} numeric vector specifying which rows can be modified: rows in the future where fix==FALSE
}

Please note that \code{target} is missing here. This is determined by the optimization routines in \code{tdmore} itself.
}
