% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataConversion.R
\name{convertDataToTdmore}
\alias{convertDataToTdmore}
\title{Convert shinyTDMore domain to TDMore domain.}
\usage{
convertDataToTdmore(model, doses, obs, covs, now)
}
\arguments{
\item{model}{model}

\item{doses}{doses}

\item{obs}{observations}

\item{covs}{covariates}

\item{now}{now date, POSIXlt date}
}
\value{
to be described
}
\description{
Convert shinyTDMore domain to TDMore domain.
}
