% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataConversion.R
\name{covsToTdmore}
\alias{covsToTdmore}
\title{Covariates conversion (shinyTDMore -> TDMore).
TODO: discuss this code within the team and test it.}
\usage{
covsToTdmore(covs, firstDoseDate)
}
\arguments{
\item{covs}{shinyTDMore covariates}

\item{firstDoseDate}{first dose date}
}
\value{
TDMore covariates
}
\description{
Covariates conversion (shinyTDMore -> TDMore).
TODO: discuss this code within the team and test it.
}
