% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newPatientDialog.R
\name{createCovariateForm}
\alias{createCovariateForm}
\title{Create the covariate form.}
\usage{
createCovariateForm(ns, input)
}
\arguments{
\item{ns}{namespace}

\item{input}{shiny input}
}
\description{
Create the covariate form.
}
