% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence_layer.R
\name{createPatient}
\alias{createPatient}
\title{Patient constructor.}
\usage{
createPatient(firstname, lastname, covariates = NULL)
}
\arguments{
\item{firstname}{patient's firstname}

\item{lastname}{patient's lastname}

\item{covariates}{named numeric with the covariates}
}
\value{
the patient
}
\description{
Patient constructor.
}
