% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelLibrary.R
\docType{data}
\name{D0_14_Troughs_1cpt_BASE_TSFD}
\alias{D0_14_Troughs_1cpt_BASE_TSFD}
\title{D0_14_Troughs_1cpt_BASE_TSFD model.}
\format{tdmore model}
\usage{
D0_14_Troughs_1cpt_BASE_TSFD
}
\description{
D0_14_Troughs_1cpt_BASE_TSFD model.
}
\keyword{datasets}
