% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelLibrary.R
\docType{data}
\name{D0_14_Troughs_1cpt_BASE_TSFD_MPC}
\alias{D0_14_Troughs_1cpt_BASE_TSFD_MPC}
\title{D0_14_Troughs_1cpt_BASE_TSFD_MPC model.}
\format{tdmore model}
\usage{
D0_14_Troughs_1cpt_BASE_TSFD_MPC
}
\description{
D0_14_Troughs_1cpt_BASE_TSFD_MPC model.
}
\keyword{datasets}
