% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelLibrary.R
\docType{data}
\name{D7_AUC_2cpt_Tlag_CYP3A5_alloWT}
\alias{D7_AUC_2cpt_Tlag_CYP3A5_alloWT}
\title{D7_AUC_2cpt_Tlag_CYP3A5_alloWT model.}
\format{tdmore model}
\usage{
D7_AUC_2cpt_Tlag_CYP3A5_alloWT
}
\description{
D7_AUC_2cpt_Tlag_CYP3A5_alloWT model.
}
\keyword{datasets}
