% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelLibrary.R
\docType{data}
\name{D7_AUC_2cpt_Tlag_CYP3A5_alloWT_MPC}
\alias{D7_AUC_2cpt_Tlag_CYP3A5_alloWT_MPC}
\title{D7_AUC_2cpt_Tlag_CYP3A5_alloWT_MPC model.}
\format{tdmore model}
\usage{
D7_AUC_2cpt_Tlag_CYP3A5_alloWT_MPC
}
\description{
D7_AUC_2cpt_Tlag_CYP3A5_alloWT_MPC model.
}
\keyword{datasets}
