% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultApp.R
\name{defaultApp}
\alias{defaultApp}
\title{Default full-featured application}
\usage{
defaultApp(
  model = tdmore::getModel(name = "pheno", dir = system.file("models", package =
    "shinytdmore")),
  now = Sys.time(),
  ...
)
}
\arguments{
\item{model}{a tdmore model with (ideally) metadata}

\item{now}{a POSIXct time representing the current time}

\item{...}{default values for the application (e.g. `regimen`, `observed`, etc)}
}
\description{
Default full-featured application
}
