% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_hotTable.R
\name{doseTable}
\alias{doseTable}
\alias{doseTableUI}
\title{Create a user interface to represent the doses for a patient}
\usage{
doseTableUI(id)

doseTable(input, output, session, state)
}
\arguments{
\item{id}{identifier for the encapsulating div}

\item{state}{`reactiveValues` object with the values `now`, `regimen` and `model`}

\item{...}{applied to the encapsulating div}
}
\description{
The interface consists of a \emph{table} (rhandsontable) that reads from the \code{state$regimen} data.frame. It has five columns:
}
\details{
\enumerate{
\item \code{time} is a POSIXct date that is represented as two seperate columns YYYY/MM/DD and HH:MM in the table
\item \code{dose} is a numeric vector. If the \code{state$model} is available, the column title is adapted.
\item \code{formulation} is a factor that shows the formulation used. If the \code{state$model} is available, the options are taken from the model. If not, this is free-form.
\item \code{fix} is a logical that shows whether the dose can be modified by the recommendation algorithm. The column is hidden from view.
}

The table has a horizontal line to reflect where \code{state$now} is situated. If \code{state$now} is empty, the line is not shown.

Edits to this dosing table are automatically synchronized back to \code{state$regimen}. The element ensures the administrations are chronological.

The user interface also features a button to \strong{add} new doses. Adding to an empty table creates a line for a dose at \strong{state$now}.
Adding to an existing table duplicates the last row. The time is shifted by either the inter-dose interval defined in the model meta-data for
the given formulation, or by \code{24} if this is not available.
}
