% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xEditableInput.R
\name{editableInput}
\alias{editableInput}
\title{Get editable input.}
\usage{
editableInput(inputId, type, value = "", options = list())
}
\arguments{
\item{inputId}{inputId}

\item{type}{type}

\item{value}{value}

\item{options}{options}
}
\value{
an editable input
}
\description{
Get editable input.
}
