% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\docType{class}
\name{FileDatabase}
\alias{FileDatabase}
\title{FileDatabase Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FileDatabase
}
\description{
FileDatabase class description
}
\details{
File database class
}
\keyword{datasets}
