% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_fitPlot.R
\name{fitPlot}
\alias{fitPlot}
\title{Fitplot creates a Population, Fit and Recommendation plot}
\usage{
fitPlot(input, output, session, state, cr = NULL)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{state}{a reactiveValues object with the state of the application. See [convertDataToTdmore()] for more information.}

\item{cr}{list of calculation-related reactives, see [calculationReactives()]}
}
\description{
Fitplot creates a Population, Fit and Recommendation plot
}
