% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_fitPlot.R
\name{fitPlotUI}
\alias{fitPlotUI}
\title{This UI element displays several plots
that show the current state of the fit.}
\usage{
fitPlotUI(id)
}
\arguments{
\item{id}{output ID}
}
\description{
It is displayed in a tabsetPanel
}
