% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo.R
\name{getAllPatients}
\alias{getAllPatients}
\title{Get all patiens from the database.}
\usage{
getAllPatients()
}
\arguments{
\item{id}{by ID}

\item{firstname}{by firstname}

\item{lastname}{by lastname}
}
\value{
the requested patient
}
\description{
Get all patiens from the database.
}
