% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeUtilities.R
\name{getDateTimeFormat}
\alias{getDateTimeFormat}
\title{Get the datetime format for this project.}
\usage{
getDateTimeFormat()
}
\arguments{
\item{posixDate}{a posix date}
}
\value{
a well formated string corresponding to the date, time zone included
}
\description{
Get the datetime format for this project.
}
