% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictionCore.R
\name{getDosingInterval}
\alias{getDosingInterval}
\title{Retrieve the dosing interval from the tdmore metadata according to the specified formulation.
If no formulation matches, dosing interval of the first formulation is returned.
If no formulation at all in metadata, 24 hours is returned.}
\usage{
getDosingInterval(model, formulation = NULL)
}
\arguments{
\item{model}{tdmore model}

\item{formulation}{formulation we are looking for, character, can be NULL}
}
\value{
a dosing interval, numeric value
}
\description{
Retrieve the dosing interval from the tdmore metadata according to the specified formulation.
If no formulation matches, dosing interval of the first formulation is returned.
If no formulation at all in metadata, 24 hours is returned.
}
