% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{getNewdata}
\alias{getNewdata}
\title{Create a newdata data.frame compatible with [tdmore::predict.tdmorefit()]}
\usage{
getNewdata(state, model)
}
\arguments{
\item{state}{reactiveValues() or list-like object with input data, see \link{shinytdmore-data} for more information}

\item{model}{a tdmore model with (ideally) metadata}
}
\value{
a tibble with at least the regimen times, troughs, and observed times, and intermediate times per 0.5h
}
\description{
Create a newdata data.frame compatible with [tdmore::predict.tdmorefit()]
}
