% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{getNewdata}
\alias{getNewdata}
\title{Create a newdata data.frame compatible with [tdmore::predict.tdmorefit()]}
\usage{
getNewdata(regimen, observed, model)
}
\arguments{
\item{regimen}{a data.frame with the following columns:
\itemize{
\item \code{time} a POSIXct time
\item \code{dose} a numeric with the dose amount
\item \code{formulation} a character vector that corresponds to formulations defined in the metadata of the model
\item \code{fix} a boolean vector describing whether the given regimen can be modified in dose recommendation
}}

\item{observed}{a data.frame with the following columns:
\itemize{
\item \code{time} a POSIXct time
\item \code{dv} the observed data
\item \code{use} whether the observation should be used in the estimation
}}

\item{model}{a tdmore model with (ideally) metadata}
}
\value{
a tibble with at least the regimen times, troughs, and observed times, and intermediate times per 0.5h
}
\description{
Create a newdata data.frame compatible with [tdmore::predict.tdmorefit()]
}
