% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataConversion.R
\name{getNewdata}
\alias{getNewdata}
\title{Get 'newdata' dataframe for tdmore predictions.}
\usage{
getNewdata(start, stop, output)
}
\arguments{
\item{start}{start time of 'TIME' column}

\item{stop}{stop time of 'TIME' column}

\item{output}{output name (e.g. 'CONC')}
}
\value{
a dataframe for tdmore
}
\description{
Get 'newdata' dataframe for tdmore predictions.
}
