% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{JsonDatabase}
\alias{JsonDatabase}
\title{JSON database class}
\description{
JsonDatabase The Json Database is an abstract superclass that makes it easy to store
data as JSON objects in a database.
}
\section{Super class}{
\code{\link[shinytdmore:Database]{shinytdmore::Database}} -> \code{JsonDatabase}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{patients}}{list of all patients}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-get}{\code{JsonDatabase$get()}}
\item \href{#method-update}{\code{JsonDatabase$update()}}
\item \href{#method-remove}{\code{JsonDatabase$remove()}}
\item \href{#method-add}{\code{JsonDatabase$add()}}
\item \href{#method-clone}{\code{JsonDatabase$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\subsection{Method \code{get()}}{
get an object with the given id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JsonDatabase$get(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{string}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the object with this id
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\subsection{Method \code{update()}}{
update an object with the given id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JsonDatabase$update(id, patient)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{string}

\item{\code{patient}}{new object details}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
this database, useful for chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove"></a>}}
\subsection{Method \code{remove()}}{
remove an object with the given id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JsonDatabase$remove(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{string}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
this database, useful for chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add"></a>}}
\subsection{Method \code{add()}}{
add the given patient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JsonDatabase$add(patient)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{patient}}{new object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the newly created ID
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JsonDatabase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
