% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_modelTab.R
\name{modelTab}
\alias{modelTab}
\title{Model tab server.}
\usage{
modelTab(
  input,
  output,
  session,
  state,
  rmdFile = system.file("model.Rmd", package = "shinytdmore")
)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{val}{main reactive container}

\item{onTabChanged}{on tab changed}
}
\description{
Model tab server.
}
