% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTab.R
\name{modelTab}
\alias{modelTab}
\title{Model tab server.}
\usage{
modelTab(input, output, session, val)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{val}{main reactive container}
}
\description{
Model tab server.
}
