% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_modelTab.R
\name{modelTabUI}
\alias{modelTabUI}
\title{Model tab user interface.}
\usage{
modelTabUI(id, dir = system.file("models", package = "tdmore"))
}
\arguments{
\item{id}{namespace id}
}
\value{
a panel
}
\description{
Model tab user interface.
}
