% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_mongodb.R
\name{MongoDatabase}
\alias{MongoDatabase}
\title{Mongo database}
\description{
Mongo database

Mongo database
}
\section{Super classes}{
\code{\link[shinytdmore:Database]{shinytdmore::Database}} -> \code{\link[shinytdmore:JsonDatabase]{shinytdmore::JsonDatabase}} -> \code{MongoDatabase}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MongoDatabase$new()}}
\item \href{#method-clone}{\code{MongoDatabase$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="shinytdmore" data-topic="JsonDatabase" data-id="add">}\href{../../shinytdmore/html/JsonDatabase.html#method-add}{\code{shinytdmore::JsonDatabase$add()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytdmore" data-topic="JsonDatabase" data-id="get">}\href{../../shinytdmore/html/JsonDatabase.html#method-get}{\code{shinytdmore::JsonDatabase$get()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytdmore" data-topic="JsonDatabase" data-id="remove">}\href{../../shinytdmore/html/JsonDatabase.html#method-remove}{\code{shinytdmore::JsonDatabase$remove()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shinytdmore" data-topic="JsonDatabase" data-id="update">}\href{../../shinytdmore/html/JsonDatabase.html#method-update}{\code{shinytdmore::JsonDatabase$update()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new mongodb connection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MongoDatabase$new(collection, db, url = "mongodb://localhost")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{collection}}{mongo collection, string}

\item{\code{db}}{database name, string}

\item{\code{url}}{connection url}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `MongoDb` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MongoDatabase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
