% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newPatientDialog.R
\name{newPatientDialog}
\alias{newPatientDialog}
\title{New patient dialog server.}
\usage{
newPatientDialog(input, output, session, onNewPatientAdded)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{onNewPatientAdded}{reactive value}
}
\description{
New patient dialog server.
}
