% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newPatientDialog.R
\name{newPatientDialogUI}
\alias{newPatientDialogUI}
\title{Create a patient form in a modal dialog.}
\usage{
newPatientDialogUI(id)
}
\arguments{
\item{id}{namespace id}
}
\description{
Create a patient form in a modal dialog.
}
