% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_hotTable.R
\name{observationTable}
\alias{observationTable}
\alias{observationTableUI}
\title{Create a user interface to represent the observations}
\usage{
observationTableUI(id)

observationTable(input, output, session, state)
}
\arguments{
\item{state}{`reactiveValues` object with at least the values `now`, `observed` and `model`}
}
\description{
The interface consists of a \emph{table} (rhandsontable) that reads from the \code{state$observed} data.frame. It has five columns:
}
\details{
\enumerate{
\item \code{date} is a POSIXct date that is represented as YYYY/MM/DD in the table
\item \code{time} is a character string HH:MM that is represented as a dropdown box
\item \code{dv} is a numeric vector. If the \code{state$model} is available, the column title is adapted.
\item \code{use} is a logical that shows whether the observation will be used by the recommendation algorithm.
}

The table has a horizontal line to reflect where \code{state$now} is situated. If \code{state$now} is empty, the line is not shown.

Edits to this observed table are automatically synchronized back to \code{state$observed}. The element ensures the rows are chronological.

The user interface also features a button to \strong{add} new observations. Adding to an empty table creates a line for an administration at \strong{state$now}.
Adding to an existing table duplicates the last row. The time is shifted by \verb{24 hours}.
}
