% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtilities.R
\name{posixToHours}
\alias{posixToHours}
\title{Convert POSIX date to hours (numeric).}
\usage{
posixToHours(posixDate)
}
\arguments{
\item{posixDate}{a date}
}
\value{
the converted dates, in hours
}
\description{
Convert POSIX date to hours (numeric).
}
