% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeUtilities.R
\name{POSIXToString}
\alias{POSIXToString}
\title{Format a POSIX datetime to a well-formatted string.}
\usage{
POSIXToString(posixDatetime)
}
\arguments{
\item{posixDatetime}{a POSIX datetime}
}
\value{
a well formated string corresponding to the date, time zone included, e.g. '2018-12-08 09:00:00 +0100'
}
\description{
Format a POSIX datetime to a well-formatted string.
}
