% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictionTab.R
\name{predictionTab}
\alias{predictionTab}
\title{Prediction tab server.}
\usage{
predictionTab(input, output, session, nsId, val)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{nsId}{namespace id}

\item{val}{main reactive container}
}
\description{
Prediction tab server.
}
