% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictionTab.R
\name{predictionTab}
\alias{predictionTab}
\title{Prediction tab server.}
\usage{
predictionTab(input, output, session, state, cr = NULL)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{state}{a reactiveValues object with the state of the application. See [convertDataToTdmore()] for more information.}

\item{cr}{list of calculation-related reactives, see [calculationReactives()]}
}
\description{
Prediction tab server.
}
