% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictionTab.R
\name{predictionTabUI}
\alias{predictionTabUI}
\title{Prediction tab user interface.}
\usage{
predictionTabUI(id)
}
\arguments{
\item{id}{namespace id}
}
\value{
a panel
}
\description{
Prediction tab user interface.
}
