% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtilities.R
\name{prepareParametersPlot}
\alias{prepareParametersPlot}
\title{Prepare the parameters plot.}
\usage{
prepareParametersPlot(populationData, individualDf, parameters)
}
\arguments{
\item{parameters}{parameters}

\item{data}{dataframe containing the prediction data}

\item{population}{logical value, true for population, false for individual prediction}
}
\description{
Prepare the parameters plot.
}
