% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictionCore.R
\name{preparePrediction}
\alias{preparePrediction}
\title{Prepare population OR individual prediction data}
\usage{
preparePrediction(doses, obs, model, covs, target, population, now)
}
\arguments{
\item{doses}{doses}

\item{obs}{observations}

\item{model}{tdmore/tdmore_mpc/tdmore_mixture model}

\item{covs}{covariates}

\item{target}{numeric vector of size 2, min and max value}

\item{population}{logical value, true for population, false for individual}

\item{now}{now date, POSIXlt date}
}
\value{
the prediction data and the tdmore data
}
\description{
Prepare population OR individual prediction data
}
