% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtilities.R
\name{preparePredictionPlot}
\alias{preparePredictionPlot}
\title{Prepare the prediction plot.}
\usage{
preparePredictionPlot(
  populationPredict,
  individualPredict,
  observed,
  target,
  model,
  now
)
}
\arguments{
\item{populationPredict}{dataframe containing the individual prediction data}

\item{individualPredict}{dataframe containing the population prediction data}

\item{observed}{a data.frame with the following columns:
\itemize{
\item \code{time} a POSIXct time
\item \code{dv} the observed data
\item \code{use} whether the observation should be used in the estimation
}}

\item{target}{a named list with min/max values to represent the target concentration}

\item{model}{a tdmore model with (ideally) metadata}

\item{now}{a POSIXct time representing the current time}
}
\description{
Prepare the prediction plot.
}
