% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtilities.R
\name{preparePredictionPlot}
\alias{preparePredictionPlot}
\title{Prepare the prediction plot.}
\usage{
preparePredictionPlot(populationDf, individualDf, obs, target, model, now)
}
\arguments{
\item{obs}{dataframe containing the observations}

\item{target}{numeric vector of size 2, min and max value}

\item{model}{tdmore model}

\item{now}{now date}

\item{data}{dataframe containing the prediction data}

\item{population}{logical value, true for population, false for individual prediction}
}
\description{
Prepare the prediction plot.
}
