% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtilities.R
\name{preparePredictionPlots}
\alias{preparePredictionPlots}
\title{Prepare population OR individual prediction plots.
This includes the prediction plot and the timeline plot.}
\usage{
preparePredictionPlots(
  populationDf,
  individualDf,
  observed,
  target,
  model,
  now,
  regimen
)
}
\arguments{
\item{target}{numeric vector of size 2, min and max value}

\item{model}{tdmore/tdmore_mpc model}

\item{now}{now date, POSIXlt date}

\item{doses}{doses}

\item{obs}{observations}

\item{covs}{covariates}
}
\value{
a list of two plots
}
\description{
Prepare population OR individual prediction plots.
This includes the prediction plot and the timeline plot.
}
