% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtilities.R
\name{preparePredictionPlots}
\alias{preparePredictionPlots}
\title{Prepare population OR individual prediction plots.
This includes the prediction plot and the timeline plot.}
\usage{
preparePredictionPlots(doses, obs, model, covs, target, population, now)
}
\arguments{
\item{doses}{doses}

\item{obs}{observations}

\item{model}{tdmore model}

\item{covs}{covariates}

\item{target}{numeric vector of size 2, min and max value}

\item{population}{logical value, true for population, false for individual}

\item{now}{now date, POSIXlt date}
}
\value{
a list of two plots
}
\description{
Prepare population OR individual prediction plots.
This includes the prediction plot and the timeline plot.
}
