% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictionCore.R
\name{prepareRecommendation}
\alias{prepareRecommendation}
\title{Prepare recommendation.}
\usage{
prepareRecommendation(doses, obs, model, covs, target, now)
}
\arguments{
\item{doses}{doses}

\item{obs}{observations}

\item{model}{tdmore model}

\item{covs}{covariates}

\item{target}{numeric vector of size 2, min and max value}

\item{now}{now date, POSIXlt date}
}
\value{
a list with pred, ipred and the recommendation
}
\description{
Prepare recommendation.
}
