% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtilities.R
\name{prepareRecommendationPlots}
\alias{prepareRecommendationPlots}
\title{Prepare recommendation plots (to be refactored).
This includes the recommendation plot and the timeline plot.}
\usage{
prepareRecommendationPlots(
  populationPredict,
  individualPredict,
  recommendationPredict,
  observed,
  target,
  model,
  now,
  regimen,
  recommendedRegimen
)
}
\arguments{
\item{populationPredict}{dataframe containing the individual prediction data}

\item{individualPredict}{dataframe containing the population prediction data}

\item{recommendationPredict}{recommendation data (contains namely ipred, ipredNew and the recommendation)}

\item{observed}{a data.frame with the following columns:
\itemize{
\item \code{time} a POSIXct time
\item \code{dv} the observed data
\item \code{use} whether the observation should be used in the estimation
}}

\item{target}{a named list with min/max values to represent the target concentration}

\item{model}{a tdmore model with (ideally) metadata}

\item{now}{a POSIXct time representing the current time}

\item{regimen}{a data.frame with the following columns:
\itemize{
\item \code{time} a POSIXct time
\item \code{dose} a numeric with the dose amount
\item \code{formulation} a character vector that corresponds to formulations defined in the metadata of the model
\item \code{fix} a boolean vector describing whether the given regimen can be modified in dose recommendation
}}

\item{recommendedRegimen}{regimen with the AMT column changed by the recommended doses}
}
\value{
a list of two plots
}
\description{
Prepare recommendation plots (to be refactored).
This includes the recommendation plot and the timeline plot.
}
