% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtilities.R
\name{prepareRecommendationPlots}
\alias{prepareRecommendationPlots}
\title{Prepare recommendation plots (to be refactored).
This includes the recommendation plot and the timeline plot.}
\usage{
prepareRecommendationPlots(doses, obs, model, covs, target, recommendation,
  now)
}
\arguments{
\item{doses}{doses}

\item{obs}{observations}

\item{model}{tdmore model}

\item{covs}{covariates}

\item{target}{numeric vector of size 2, min and max value}

\item{recommendation}{recommendation (contains pred, ipred and the recommendation)}

\item{now}{now date}
}
\value{
a list of two plots
}
\description{
Prepare recommendation plots (to be refactored).
This includes the recommendation plot and the timeline plot.
}
