% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtilities.R
\name{prepareTimelinePlot}
\alias{prepareTimelinePlot}
\title{Prepare the timeline.}
\usage{
prepareTimelinePlot(regimen, model, now)
}
\arguments{
\item{model}{tdmore model}

\item{now}{now date, currently not used here}

\item{doses}{dataframe containing the doses}
}
\description{
Prepare the timeline.
}
