% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtilities.R
\name{prepareTimelinePlot}
\alias{prepareTimelinePlot}
\title{Prepare the timeline.}
\usage{
prepareTimelinePlot(doses, xlim, model, now)
}
\arguments{
\item{doses}{dataframe containing the doses}

\item{xlim}{ggplot xlim argument}

\item{model}{tdmore model}

\item{now}{now date, currently not used here}
}
\description{
Prepare the timeline.
}
