% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_hotTable.R
\name{recommendationTable}
\alias{recommendationTable}
\title{The recommendationTable uses the state$regimen and state$recommendation tables}
\usage{
recommendationTable(input, output, session, state)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{state}{a reactiveValues object with the state of the application. See [convertDataToTdmore()] for more information.}
}
\description{
The recommendationTable uses the state$regimen and state$recommendation tables
}
