% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientsTab.R
\name{removePatientConfirmationDialog}
\alias{removePatientConfirmationDialog}
\title{Remove patient confirmation dialog.}
\usage{
removePatientConfirmationDialog(patientName)
}
\arguments{
\item{patientName}{the patient name}
}
\value{
a modal dialog
}
\description{
Remove patient confirmation dialog.
}
