% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientsTab.R
\name{renderPatientTable}
\alias{renderPatientTable}
\title{Render the patients table.}
\usage{
renderPatientTable(input, ns, db)
}
\arguments{
\item{input}{the shiny input object}

\item{ns}{namespace}

\item{db}{database}
}
\description{
Render the patients table.
}
