% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveProject.R
\name{saveProject}
\alias{saveProject}
\title{Save project server}
\usage{
saveProject(input, output, session, onTabChanged, val, db)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{onTabChanged}{onTabChanged reactive values}

\item{val}{main reactive container}

\item{db}{database}
}
\description{
Save project server
}
