% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataConversion.R
\name{shinytdmore-data}
\alias{shinytdmore-data}
\alias{defaultData}
\alias{convertDataToTdmore}
\title{Data format description}
\arguments{
\item{state}{reactiveValues() or list-like object with input data, see \link{shinytdmore-data} for more information}

\item{force}{if FALSE, emits a silent error (see [shiny::req()]) when no shiny data is available. If TRUE, tries to generate the args anyway.}

\item{model}{a tdmore model with (ideally) metadata}

\item{regimen}{a data.frame with the following columns:
\itemize{
\item \code{time} a POSIXct time
\item \code{dose} a numeric with the dose amount
\item \code{formulation} a character vector that corresponds to formulations defined in the metadata of the model
\item \code{fix} a boolean vector describing whether the given regimen can be modified in dose recommendation
}}

\item{observed}{a data.frame with the following columns:
\itemize{
\item \code{time} a POSIXct time
\item \code{dv} the observed data
\item \code{use} whether the observation should be used in the estimation
}}

\item{covariates}{a data.frame with the following columns:
\itemize{
\item \code{time} a POSIXct time
\item \code{...} columns corresponding to covariates required in the tdmore model. If columns are missing, the method returns an error.
}}

\item{now}{a POSIXct time representing the current time}

\item{target}{a named list with min/max values to represent the target concentration}
}
\description{
Data format description
}
\details{
Shinytdmore uses the following data objects in a reactiveValues() state:
\itemize{
\item \code{state$model} a tdmore model with (ideally) metadata
\item \code{state$regimen} a data.frame with the following columns:
\itemize{
\item \code{time} a POSIXct time
\item \code{dose} a numeric with the dose amount
\item \code{formulation} a character vector that corresponds to formulations defined in the metadata of the model
\item \code{fix} a boolean vector describing whether the given regimen can be modified in dose recommendation
}
\item \code{state$observed} a data.frame with the following columns:
\itemize{
\item \code{time} a POSIXct time
\item \code{dv} the observed data
\item \code{use} whether the observation should be used in the estimation
}
\item \code{state$covariates} a data.frame with the following columns:
\itemize{
\item \code{time} a POSIXct time
\item \code{...} columns corresponding to covariates required in the tdmore model. If columns are missing, the method returns an error.
}
\item \code{state$now} a POSIXct time representing the current time
}

\code{\link[=convertDataToTdmore]{convertDataToTdmore()}} converts this into a list of arguments compatible with \link[tdmore:estimate]{tdmore::estimate}, \link[tdmore:predict.tdmore]{tdmore::predict.tdmore} and \link[tdmore:findDose]{tdmore::findDose}.
\itemize{
\item \code{model} the tdmore model
\item \code{regimen} all treatments in the right format
\item \code{observed} all observed values where use=TRUE, and with the \code{dv} column renamed to the default model output
\item \code{covariates} all covariates, with an added covariate \code{FORM} that contains the formulation
\item \code{t0} a POSIXct time corresponding to time \code{0}
\item \code{now} a numeric time, corresponding to the different between t0 and now
\item \code{doseRows} numeric vector specifying which rows can be modified: rows in the future where fix==FALSE
}

Please note that \code{target} is missing here. This is determined by the optimization routines in \code{tdmore} itself.
}
\section{Functions}{
\itemize{
\item \code{defaultData}: Default data

\item \code{convertDataToTdmore}: Convert shinytdmore data to tdmore format
}}

