% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTdmore.R
\name{shinyTdmore}
\alias{shinyTdmore}
\title{Server component of Shiny TDMore.}
\usage{
shinyTdmore(input, output, session, modules, db)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{db}{database, uses in-memory database if missing}

\item{conf}{named list with modules}
}
\value{
server component
}
\description{
Server component of Shiny TDMore.
}
