% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTdmore.R
\name{shinyTdmore}
\alias{shinyTdmore}
\title{Get server component of Shiny TDMore.}
\usage{
shinyTdmore(input, output, session, conf)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{conf}{modules configuration}
}
\value{
server component
}
\description{
Get server component of Shiny TDMore.
}
