% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveProject.R
\name{tabHasChanged}
\alias{tabHasChanged}
\title{Say if the user went out of the 'Prediction' tab.}
\usage{
tabHasChanged(input, val)
}
\arguments{
\item{input}{shiny input object}

\item{val}{main reactive container}
}
\value{
a logical value
}
\description{
Say if the user went out of the 'Prediction' tab.
}
