% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_target.R
\name{targetInput}
\alias{targetInput}
\alias{targetInputUI}
\title{Input element representing the target upper limit and lower limits
The input element is synchronized to state$target$min and state$target$max
If state$target does not exist, it creates this as list(min=NA, max=NA).
NA assumes the model-defined min/max criteria will be used.}
\usage{
targetInputUI(id)

targetInput(input, output, session, state)
}
\arguments{
\item{id}{unique identifier}

\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{state}{a reactiveValues object with the state of the application. See [convertDataToTdmore()] for more information.}
}
\description{
Input element representing the target upper limit and lower limits
The input element is synchronized to state$target$min and state$target$max
If state$target does not exist, it creates this as list(min=NA, max=NA).
NA assumes the model-defined min/max criteria will be used.
}
