% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientModel.R
\name{updatePatientModel}
\alias{updatePatientModel}
\title{Update patient model and covariates}
\usage{
updatePatientModel(patient, modelName)
}
\arguments{
\item{patient}{the given patient}

\item{modelName}{model name}
}
\value{
the patient
}
\description{
Update patient model and covariates
}
