% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtilities.R
\name{updatePlot}
\alias{updatePlot}
\title{Update plot using animation. NOT USED.}
\usage{
updatePlot(plot, outputId)
}
\arguments{
\item{plot}{the given plot}

\item{outputId}{the output ID}
}
\description{
Update plot using animation. NOT USED.
}
